+++ src/Adminx/Views/NoBaseViewResponse.php
@@ -1,0 +1,20 @@
+<?php
+
+/*
+ * This file is part of Adminx.
+ *
+ * Copyright 2020-2022 Parsa Shahmaleki <parsampsh@gmail.com>
+ *
+ * Adminx project is Licensed Under MIT.
+ * For more information, please see the LICENSE file.
+ */
+
+namespace Adminx\Views;
+
+class NoBaseViewResponse
+{
+    function __construct(
+        public $response
+    )
+    {}
+}