+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Changed naming style of a lot of function and variables
 - Some changes in structure of the plugin system
 - Dropped support for PHP 7.4
+- Added `Adminx\Views\NoBaseViewResponse` class for custom pages to return straight response without puting it in the base view
 
 ## 0.1.1
 - Just some improvements